ml_service/requirements.txt
flask==3.0.0
numpy
Pillow
tensorflow-cpu
